NET-59: move suite to separate file
